// packages/web/src/shared/api/generated/model/analyticsSummaryMetricDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummaryMetricDto {
  totalInBaseCurrency: string;
  previousTotalInBaseCurrency: string;
  /** @nullable */
  percentageChange?: number | null;
  transactionCount: number;
}